update to dojo 1.4; use tundra theme (dojo)
So.  Here we switch to dojo 1.4 (and actually start using it for something);
we move away from our hand-generated table and to a dojo grid; as well as some
primitive layout and "package group selection".  There are a number of updates
to the REST controller to support this, as well as tt2, etc.

There's still a ton of cleanup to be done (not the least of which is the icon
sizes in the grid!), but we're getting somewhere at least :)